requirements.txt
cassandra-driver==3.25.0
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
imbalanced-learn==0.10.1

